pi/requirements.txt
python-dotenv
requests
watchdog